// src/pages/siteinspection/types/siteinspections.type.ts
export interface Question {
  id: string
  text: string
  options: {
    id: string
    label: string
  }[]
}

export interface RatingQuestion {
  id: string
  text: string
}

export interface RatingOption {
  id: string
  label: string
  value: number
}

export interface RespondentInfo {
  clientType: string
  date: string
  sex: string
  age: string
  contactNumber: string
  siteInspections: string
}

export interface SurveyState {
  selectedAnswers: { [key: string]: string }
  selectedRatings: { [key: string]: string }
  respondentInfo: RespondentInfo
  comments: string
}